+++ HISTORY.rst
@@ -13,6 +13,7 @@
 - Make it clear when a setting is being overrided by an environment
   variable. Add better text explanation. Change the column to a check
   mark widget.
+- Add icons to the smart settings links.
 
 3.3.3 (2019-12-05)
 ==================